requirements.txt
gooddata-sdk ~= 1.20.0
googletrans  ~= 4.0.0rc1
python-dotenv ~= 1.0.1